+++ novelty-books/src/components/Purchases/Purchases.jsx
@@ -1,6 +1,6 @@
 "use client"
 import React from "react";
-import { useStateContext } from "../../context/StateContext"
+import { useStateContext } from "../../../context/StateContext"
 import { useRouter } from "next/navigation";
 import Link from "next/link";
 import styles from "./Purchases.module.css"
